Remove lerna
