--- modules/box/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "box-styled",
-  "version": "2.0.0-0",
-  "description": "React layout component built with styled-components",
-  "main": "dist/Box.js",
-  "scripts": {
-    "prepublish": "babel src -d dist",
-    "test": "ava"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "grid-styled-constants": "2.0.0-0",
-    "styled-components": "^2.0.1"
-  },
-  "devDependencies": {
-    "ava": "^0.19.1",
-    "babel-core": "^6.24.1",
-    "babel-preset-env": "^1.5.2",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "react": "^15.5.4",
-    "react-test-renderer": "^15.5.4"
-  },
-  "ava": {
-    "require": [
-      "babel-register"
-    ],
-    "babel": "inherit"
-  }
-}
