--- modules/constants/README.md
@@ -1,4 +1,0 @@
-
-# grid-styled-constants
-
-Default style constants for grid-styled
