--- modules/constants/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "grid-styled-constants",
-  "version": "2.0.0-0",
-  "description": "Default style constants for grid-styled",
-  "main": "index.js",
-  "scripts": {},
-  "keywords": [],
-  "author": "Brent Jackson",
-  "license": "MIT"
-}
