--- modules/core/src/index.js
@@ -1,23 +1,0 @@
-import Box from 'box-styled'
-import Flex from 'flex-styled'
-import Grid from 'grid-styled-grid'
-
-import Golden from './Golden'
-import { Half, Third, Quarter } from './Fractions'
-
-export { default as Box } from 'box-styled'
-export { default as Flex } from 'flex-styled'
-export { default as Grid } from 'grid-styled-grid'
-
-export { default as Golden } from './Golden'
-export { Half, Third, Quarter } from './Fractions'
-
-export default {
-  Box,
-  Flex,
-  Grid,
-  Golden,
-  Half,
-  Third,
-  Quarter
-}
