--- modules/flex/.babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    "env",
-    "stage-0",
-    "react"
-  ]
-}
