--- modules/flex/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "flex-styled",
-  "version": "2.0.0-0",
-  "description": "Responsive React flexbox component built with styled-components",
-  "main": "dist/Flex.js",
-  "scripts": {
-    "prepublish": "babel src -d dist",
-    "test": "ava"
-  },
-  "dependencies": {
-    "box-styled": "2.0.0-0",
-    "grid-styled-constants": "2.0.0-0",
-    "styled-components": "^2.0.1"
-  },
-  "keywords": [],
-  "author": "Brent Jackson",
-  "license": "MIT",
-  "devDependencies": {
-    "ava": "^0.19.1",
-    "babel-preset-env": "^1.5.2",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "react": "^15.5.4",
-    "react-test-renderer": "^15.5.4"
-  },
-  "ava": {
-    "require": [
-      "babel-register"
-    ],
-    "babel": "inherit"
-  }
-}
