+++ package.json
@@ -1,38 +1,37 @@
 {
   "name": "grid-styled",
-  "private": true,
+  "version": "2.0.0-0",
   "description": "Responsive React grid system built with styled-components",
   "scripts": {
-    "bootstrap": "lerna bootstrap",
-    "_lint": "stylelint ./src/**/*.js",
-    "test": "lerna run test",
+    "prepublish": "babel src -d dist",
+    "test": "ava",
     "start": "webpack-dev-server"
   },
-  "keywords": [],
+  "keywords": [
+    "react",
+    "styled-components",
+    "grid",
+    "flexbox",
+    "layout",
+    "css"
+  ],
   "author": "Brent Jackson",
   "license": "MIT",
   "devDependencies": {
     "ava": "^0.19.1",
     "babel-cli": "^6.16.0",
     "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-env": "^1.5.2",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-0": "^6.16.0",
     "babel-register": "^6.16.3",
-    "classnames": "^2.2.5",
-    "enzyme": "^2.5.1",
     "html-webpack-plugin": "^2.24.0",
-    "lerna": "^2.0.0-rc.5",
     "react": "^15.5.4",
     "react-addons-test-utils": "^15.3.2",
     "react-dom": "^15.5.4",
     "react-live": "^1.5.4",
+    "react-test-renderer": "^15.5.4",
     "react-x-ray": "^1.0.0-1",
-    "rebass": "^0.3.3",
-    "recompose": "^0.20.2",
-    "stylelint": "^7.5.0",
-    "stylelint-config-standard": "^14.0.0",
-    "stylelint-processor-styled-components": "0.0.3",
     "webpack": "^2.6.1",
     "webpack-dev-server": "^2.4.5"
   },
