+++ src/Half.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Grid from './Grid'
+
+const Half = props => <Grid {...props} width={[ 1, 1/2 ]} />
+
+export default Half
