+++ src/Quarter.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Grid from './Grid'
+
+const Quarter = props => <Grid {...props} width={[ 1, 1/4 ]} />
+
+export default Quarter
