+++ src/Third.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Grid from './Grid'
+
+const Third = props => <Grid {...props} width={[ 1, 1/3 ]} />
+
+export default Third
