+++ src/constants.js
@@ -1,10 +1,10 @@
-const breakpoints = [
+export const breakpoints = [
   40,
   52,
   64
 ]
 
-const space = [
+export const space = [
   0,
   8,
   16,
@@ -12,7 +12,7 @@ const space = [
   64
 ]
 
-module.exports = {
+export default {
   breakpoints,
   space
 }
