+++ webpack.config.js
@@ -12,8 +12,7 @@ module.exports = {
 
   resolve: {
     alias: {
-      'grid-styled': path.join(__dirname, 'modules/core/src/index.js'),
-      'box-styled': path.join(__dirname, 'modules/box/src/Box.js'),
+      'grid-styled': path.join(__dirname, 'src/index.js')
     }
   },
